Update build task to generate .json file. Build task takes path prefix optional parameter. Implement merging of .json-files in wwwroot/slides folder. Start slide is handled by adding tag "START" to the slide. Update SlidesVerifier to handle all those changes. Current Slide is now handled in GameState. Split up Slides.json into 3 files, mostly for testing but it's nice so keep it I guess.
