+++ Framework/wwwroot/slidefiles.json
@@ -1,0 +1,1 @@
+["slides/hm305.json","slides/hm3_hallway.json","slides/mics_minigames.json"]
