+++ Framework/wwwroot/slides/hm305.json
@@ -1,0 +1,183 @@
+{
+    "HM305": {
+        "Tags": ["START"],
+        "Image": "images/HM305_fromEntrance.jpg",
+        "Buttons": {
+            "Blackboard": {
+                "Type": "polygon",
+                "Points": "1085,285 1085,440 1282,441 1288,254",
+                "Actions": [["Route", "HM305blackboard"]]
+            },
+            "Beamer": {
+                "Type": "polygon",
+                "Points": "915,197,956,217,956,248,918,256,868,244,869,217",
+                "Actions": [
+                    ["RequireItem", "goldkey"],
+                    ["Route", "HM305Beamer"]
+                ]
+            },
+            "Back": {
+                "Type": "image",
+                "Image": "UI_Images/backImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "BackPolygon": {
+                "Type": "polygon",
+                "Points": "27,15 95,17 102,29 102,95 90,104 27,104 17,95 15,27",
+                "Actions": [["Route", "HM305DoorClosed"]]
+            },
+            "RectTest": {
+                "Type": "rect",
+                "Points": "100,100,200,200",
+                "Actions": [
+                    ["StartBlock", "T"],
+                    ["RequireGameState", "!HM305.CircleTest"],
+                    ["SetGameState", "HM305.CircleTest", "true"],
+                    ["Exit", ""],
+                    ["EndBlock", "T"],
+                    ["SetGameState", "HM305.CircleTest", "false"]
+                ],
+                "Visible": true
+            },
+            "CircleTest": {
+                "Type": "circle",
+                "Points": "100,100,50",
+                "Actions": [
+                    ["Route", "MinigameTest3"],
+                    ["StopMusic", ""]
+                ],
+                "Visible": true
+            },
+            "PresetTest": {
+                "Type": "preset",
+                "Image": "right",
+                "Actions": [["PlayMusic", "audio/doom-soundtrack.wav"]]
+            }
+        }
+    },
+
+    "HM305DoorClosed": {
+        "Image": "images/HM305_DoorClose.jpg",
+        "Buttons": {
+            "HM305DoorClosed": {
+                "Type": "polygon",
+                "Points": "1206,155 861,160 856,878 1235,876",
+                "Actions": [
+                    ["RequireItem", "goldkey"],
+                    ["Route", "HM305DoorOpen"]
+                ]
+            },
+            "Back": {
+                "Type": "image",
+                "Image": "UI_Images/backImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "BackPolygon": {
+                "Type": "polygon",
+                "Points": "27,15 95,17 102,29 102,95 90,104 27,104 17,95 15,27",
+                "Actions": [["Route", "HM305"]]
+            },
+            "CodeTerminal": {
+                "Type": "polygon",
+                "Points": "764,458 764,491 775,495 775,465 764,458 792,39",
+                "Actions": [["Route", "HM305CodeTerminal"]],
+                "Visible": true
+            }
+        }
+    },
+
+    "HM305CodeTerminal": {
+        "Type": "Minigame",
+        "FallbackSlide": "HM305DoorClosed",
+        "MinigameDefClassName": "Framework.Minigames.CodeTerminal"
+    },
+
+    "HM305DoorOpen": {
+        "Image": "images/HM305_DoorOpen.jpg",
+        "Buttons": {
+            "HM305DoorOpened": {
+                "Type": "polygon",
+                "Points": "1194,155 849,160 844,878 1223,876 ",
+                "Actions": [["Route", "HM3_hallwayW"]]
+            },
+            "Back": {
+                "Type": "image",
+                "Image": "UI_Images/backImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "BackPolygon": {
+                "Type": "polygon",
+                "Points": "27,15 95,17 102,29 102,95 90,104 27,104 17,95 15,27",
+                "Actions": [["Route", "HM305"]]
+            }
+        }
+    },
+
+    "HM305Beamer": {
+        "Image": "images/HM305_beamer.jpg",
+        "Buttons": {
+            "key": {
+                "Type": "image",
+                "Points": "780,460,100,100",
+                "Actions": [],
+                "Image": "InventoryImages/key.png"
+            },
+            "key_polygon": {
+                "Type": "polygon",
+                "Points": "878,502 871,514 866,517 864,512 856,519 849,512 842,517 837,514 830,514 830,522 825,519 818,529 810,531 801,531 793,531 791,529 788,526 786,524 784,522 784,517 781,512 782,495 793,490 808,486 825,502 827,497 830,502",
+                "Actions": [["Route", "MinigameTest"]]
+            },
+            "Back": {
+                "Type": "image",
+                "Image": "UI_Images/backImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "BackPolygon": {
+                "Type": "polygon",
+                "Points": "27,15 95,17 102,29 102,95 90,104 27,104 17,95 15,27",
+                "Actions": [["Route", "HM305"]]
+            },
+            "right": {
+                "Type": "preset",
+                "Image": "right",
+                "Actions": []
+            },
+            "left": {
+                "Type": "preset",
+                "Image": "left",
+                "Actions": []
+            },
+            "top": {
+                "Type": "preset",
+                "Image": "top",
+                "Actions": []
+            },
+            "bottom": {
+                "Type": "preset",
+                "Image": "bottom",
+                "Actions": []
+            }
+        }
+    },
+
+    "HM305blackboard": {
+        "Image": "images/HM305_blackboard.jpg",
+        "Buttons": {
+            "Back": {
+                "Type": "image",
+                "Image": "UI_Images/backImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "BackPolygon": {
+                "Type": "polygon",
+                "Points": "27,15 95,17 102,29 102,95 90,104 27,104 17,95 15,27",
+                "Actions": [["Route", "HM305"]]
+            }
+        }
+    }
+}
