+++ Framework/wwwroot/slides/hm3_hallway.json
@@ -1,0 +1,180 @@
+{
+    "HM3_hallwayW": {
+        "Image": "images/HM3_hallwayW.jpg",
+        "Buttons": {
+            "LeftPreset": {
+                "Type": "preset",
+                "Image": "left",
+                "Actions": [["Route", "HM3_hallwayS"]]
+            },
+            "RightPreset": {
+                "Type": "preset",
+                "Image": "right",
+                "Actions": [["Route", "HM3_hallwayN"]]
+            },
+            "Left": {
+                "Type": "image",
+                "Image": "UI_Images/leftImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "Right": {
+                "Type": "image",
+                "Image": "UI_Images/rightImg.png",
+                "Points": "10,10,3700,100",
+                "Actions": []
+            },
+            "RightPolygon": {
+                "Type": "polygon",
+                "Points": "1856,12 1875,17 1890,26 1900,41 1907,62 1902,84 1890,97 1873,108 1852,109 1830,101 1815,82 1810,62 1815,41 1830,24",
+                "Actions": [["Route", "HM3_hallwayN"]]
+            },
+            "LeftPolygon": {
+                "Type": "polygon",
+                "Points": "58,14 77,19 92,27 103,43 109,63 104,86 92,99 75,109 55,111 32,103 17,84 12,63 17,43 32,26 ",
+                "Actions": [["Route", "HM3_hallwayS"]]
+            }
+        }
+    },
+
+    "HM3_hallwayN": {
+        "Image": "images/HM3_hallwayN.jpg",
+        "Buttons": {
+            "LeftPreset": {
+                "Type": "preset",
+                "Image": "left",
+                "Actions": [["Route", "HM3_hallwayW"]]
+            },
+            "RightPreset": {
+                "Type": "preset",
+                "Image": "right",
+                "Actions": [["Route", "HM3_hallwayE"]]
+            },
+            "Left": {
+                "Type": "image",
+                "Image": "UI_Images/leftImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "Right": {
+                "Type": "image",
+                "Image": "UI_Images/rightImg.png",
+                "Points": "10,10,3700,100",
+                "Actions": []
+            },
+            "RightPolygon": {
+                "Type": "polygon",
+                "Points": "1856,12 1875,17 1890,26 1900,41 1907,62 1902,84 1890,97 1873,108 1852,109 1830,101 1815,82 1810,62 1815,41 1830,24",
+                "Actions": [["Route", "HM3_hallwayE"]]
+            },
+            "LeftPolygon": {
+                "Type": "polygon",
+                "Points": "58,14 77,19 92,27 103,43 109,63 104,86 92,99 75,109 55,111 32,103 17,84 12,63 17,43 32,26 ",
+                "Actions": [["Route", "HM3_hallwayW"]]
+            }
+        }
+    },
+
+    "HM3_hallwayE": {
+        "Image": "images/HM3_hallwayE.jpg",
+        "Buttons": {
+            "LeftPreset": {
+                "Type": "preset",
+                "Image": "left",
+                "Actions": [["Route", "HM3_hallwayN"]]
+            },
+            "RightPreset": {
+                "Type": "preset",
+                "Image": "right",
+                "Actions": [["Route", "HM3_hallwayS"]]
+            },
+            "Left": {
+                "Type": "image",
+                "Image": "UI_Images/leftImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "Right": {
+                "Type": "image",
+                "Image": "UI_Images/rightImg.png",
+                "Points": "10,10,3700,100",
+                "Actions": []
+            },
+            "RightPolygon": {
+                "Type": "polygon",
+                "Points": "1856,12 1875,17 1890,26 1900,41 1907,62 1902,84 1890,97 1873,108 1852,109 1830,101 1815,82 1810,62 1815,41 1830,24",
+                "Actions": [["Route", "HM3_hallwayS"]]
+            },
+            "LeftPolygon": {
+                "Type": "polygon",
+                "Points": "58,14 77,19 92,27 103,43 109,63 104,86 92,99 75,109 55,111 32,103 17,84 12,63 17,43 32,26 ",
+                "Actions": [["Route", "HM3_hallwayN"]]
+            },
+            "HM305": {
+                "Type": "polygon",
+                "Points": "237,441 256,884 479,888 467,436",
+                "Actions": [["Route", "HM305"]]
+            }
+        }
+    },
+
+    "HM3_hallwayS": {
+        "Image": "images/HM3_hallwayS.jpg",
+        "Buttons": {
+            "LeftPreset": {
+                "Type": "preset",
+                "Image": "left",
+                "Actions": [["Route", "HM3_hallwayE"]]
+            },
+            "RightPreset": {
+                "Type": "preset",
+                "Image": "right",
+                "Actions": [["Route", "HM3_hallwayW"]]
+            },
+            "Left": {
+                "Type": "image",
+                "Image": "UI_Images/leftImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "Right": {
+                "Type": "image",
+                "Image": "UI_Images/rightImg.png",
+                "Points": "10,10,3700,100",
+                "Actions": []
+            },
+            "RightPolygon": {
+                "Type": "polygon",
+                "Points": "1856,12 1875,17 1890,26 1900,41 1907,62 1902,84 1890,97 1873,108 1852,109 1830,101 1815,82 1810,62 1815,41 1830,24",
+                "Actions": [["Route", "HM3_hallwayW"]]
+            },
+            "LeftPolygon": {
+                "Type": "polygon",
+                "Points": "58,14 77,19 92,27 103,43 109,63 104,86 92,99 75,109 55,111 32,103 17,84 12,63 17,43 32,26 ",
+                "Actions": [["Route", "HM3_hallwayE"]]
+            },
+            "Locker": {
+                "Type": "polygon",
+                "Points": "38,558 38,676 135,664 135,558",
+                "Actions": [["Route", "HM3_Locker"]]
+            }
+        }
+    },
+
+    "HM3_Locker": {
+        "Image": "images/HM3_lockerOpened.jpg",
+        "Buttons": {
+            "Back": {
+                "Type": "image",
+                "Image": "UI_Images/backImg.png",
+                "Points": "10,10,100,100",
+                "Actions": []
+            },
+            "BackPolygon": {
+                "Type": "polygon",
+                "Points": "27,15 95,17 102,29 102,95 90,104 27,104 17,95 15,27",
+                "Actions": [["Route", "HM3_hallwayS"]]
+            }
+        }
+    }
+}
